arm64 warnings eliminated.

+++ PagerView/PagerItemView.m
@@ -30,7 +30,7 @@ typedef enum {
 }
 
 - (NSString*)description {
-	return [NSString stringWithFormat:@"[%@. DataSource: %d]", NSStringFromClass(self.class), self.datasourceIndex];
+	return [NSString stringWithFormat:@"[%@. DataSource: %lu]", NSStringFromClass(self.class), (unsigned long)self.datasourceIndex];
 }
 
 - (void)touchesBegan:(NSSet *)touches withEvent:(UIEvent *)event {
